# contracts/openapi/openapi.json
{
  "components" : {
    "schemas" : {
      "AdminSessionResponse" : {
        "properties" : {
          "accessExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "accessToken" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserView"
          }
        },
        "type" : "object"
      },
      "AuthResult" : {
        "properties" : {
          "accessExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "accessToken" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserView"
          }
        },
        "type" : "object"
      },
      "ContentRequest" : {
        "properties" : {
          "body" : {
            "minLength" : 1,
            "type" : "string"
          },
          "category" : {
            "enum" : [ "RECOMMENDED", "LOCAL", "ACTIVITY", "STORE" ],
            "type" : "string"
          },
          "mediaId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "summary" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 160,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "body", "category", "summary", "title" ],
        "type" : "object"
      },
      "ContentView" : {
        "properties" : {
          "body" : {
            "type" : "string"
          },
          "category" : {
            "enum" : [ "RECOMMENDED", "LOCAL", "ACTIVITY", "STORE" ],
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "followed" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "media" : {
            "$ref" : "#/components/schemas/SignedMedia"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "PUBLISHED", "UNPUBLISHED" ],
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CredentialsRequest" : {
        "properties" : {
          "password" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          },
          "username" : {
            "maxLength" : 64,
            "minLength" : 3,
            "pattern" : "[A-Za-z0-9_]+",
            "type" : "string"
          }
        },
        "required" : [ "password", "username" ],
        "type" : "object"
      },
...
      "MediaResponse" : {
        "properties" : {
          "contentType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "originalFilename" : {
            "type" : "string"
          },
          "sha256" : {
            "type" : "string"
          },
          "sizeBytes" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PageResponseContentView" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ContentView"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageResponseMediaResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/MediaResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RefreshRequest" : {
        "properties" : {
          "refreshToken" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
...
      "SignedMedia" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserView" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "role" : {
            "enum" : [ "USER", "ADMIN_VIEWER", "ADMIN" ],
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "adminRefreshCookie" : {
        "in" : "cookie",
        "name" : "phonecar_admin_refresh",
        "type" : "apiKey"
      },
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "PhoneCar API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/admin/discovery/contents" : {
      "get" : {
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseContentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-content-controller" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "admin-content-controller" ]
      }
    },
    "/api/v1/admin/discovery/contents/{id}" : {
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "admin-content-controller" ]
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-content-controller" ]
      }
    },
    "/api/v1/admin/discovery/contents/{id}/publish" : {
      "post" : {
        "operationId" : "publish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-content-controller" ]
      }
    },
    "/api/v1/admin/discovery/contents/{id}/unpublish" : {
      "post" : {
        "operationId" : "unpublish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-content-controller" ]
      }
    },
    "/api/v1/admin/media" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseMediaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-media-controller" ]
      },
      "post" : {
        "operationId" : "upload",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "admin-media-controller" ]
      }
    },
    "/api/v1/admin/media/{id}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "admin-media-controller" ]
      }
    },
    "/api/v1/auth/admin/login" : {
      "post" : {
        "operationId" : "login_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CredentialsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSessionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ ],
        "tags" : [ "admin-session-controller" ]
      }
    },
    "/api/v1/auth/admin/logout" : {
      "post" : {
        "operationId" : "logout_1",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "adminRefreshCookie" : [ ]
        } ],
        "tags" : [ "admin-session-controller" ]
      }
    },
    "/api/v1/auth/admin/refresh" : {
      "post" : {
        "operationId" : "refresh_1",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSessionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "adminRefreshCookie" : [ ]
        } ],
        "tags" : [ "admin-session-controller" ]
      }
    },
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CredentialsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ ],
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "operationId" : "logout",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ ],
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ ],
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/register" : {
      "post" : {
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CredentialsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResult"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ ],
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/discovery/contents/{id}" : {
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "discovery-content-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}